requirements.txt
jupyter
pandas
numpy
matplotlib
seaborn
black
isort
flake8
pylint
ipywidgets
python-dotenv
openai
